chore(Algebra/Ring/Int): split `Algebra/Ring/Int.lean` into `Defs.lean` and `Basic.lean`
The definition of the ring structure on the integers in this file was accompanied by a set of miscellaneous lemmas. If we split off the lemmas to their own file, they are not needed nearly as often.

(Hopefully this also allows some cleaning up around `Data.Int.Cast` in the future, although that's still a bit too much of a mess right now.)